+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 # Change Log
 All notable changes to COMMIT will be documented in this file.
 
+## [2.0.0] - 2023-##-##
+
+### Changed
+- Default `ndirs=500` in `core.generate_kernels()` and in `trk2dictionary.run()`
+- Expire the deprecated `ndirs` parameter in `amico.core.setup()`
+- Expire the deprecated `filename_trk` and `gen_trk` parameters in `trk2dictionary.run()`
+
 ## [1.6.4] - 2023-02-14
 
 ### Fixed